% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_source}
\alias{data_source}
\title{Data Source Information}
\description{
This is a dataframe containing the information about all the raw data in this package
}
\examples{
data(data_source)
summary(data_source)
}
\keyword{datasets}
