% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog712package-package.R
\docType{package}
\name{geog712package-package}
\alias{geog712package}
\alias{geog712package-package}
\title{geog712package: GEOG 712 Package}
\description{
This package is an activity completed by Zehui Yin for the course GEOG 712 Reproducible Research Workflow with GitHub and R, taught by Dr. Antonio Paez in Fall 2024.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zehuiyin/geog712package}
  \item \url{https://zehuiyin.github.io/geog712package/}
  \item Report bugs at \url{https://github.com/zehuiyin/geog712package/issues}
}

}
\author{
\strong{Maintainer}: Zehui Yin \email{yinz39@mcmaster.ca} (\href{https://orcid.org/0000-0001-6954-7918}{ORCID})

}
\keyword{internal}
