% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grocery_CT}
\alias{grocery_CT}
\title{Grocery Stores in Census Tracts in Hamilton}
\format{
A \emph{sf} dataframe with 4 columns:
\describe{
  \item{GEOUID}{The unique ID for the geographical unit}
  \item{POP21}{The population count in the census tract from 2021 Canadian Census}
  \item{Grocery}{The number of grocery stores in that census tract}
  \item{geometry}{Geometry column for \emph{sf} dataframe}
}
}
\description{
This is a \emph{sf} dataframe containing the information about the number of grocery
stores in each Census Tract in Hamilton
}
\examples{
data(grocery_CT)
summary(grocery_CT[,c("GEOUID", "POP21", "Grocery")])
head(grocery_CT[,c("GEOUID", "POP21", "Grocery")])
}
\keyword{datasets}
