% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grocery_store_hamilton_data-package.R
\docType{package}
\name{GroceryStoreHamiltonData-package}
\alias{GroceryStoreHamiltonData}
\alias{GroceryStoreHamiltonData-package}
\title{GroceryStoreHamiltonData: Data for the paper "Food Deserts or Food Oases? Predicting Grocery Store Locations in Hamilton, Ontario"}
\description{
This package is an activity completed by Zehui Yin for the course GEOG 712 Reproducible Research Workflow with GitHub and R, taught by Dr. Antonio Paez in Fall 2024.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zehuiyin/GroceryStoreHamiltonData}
  \item \url{https://zehuiyin.github.io/GroceryStoreHamiltonData/}
  \item Report bugs at \url{https://github.com/zehuiyin/GroceryStoreHamiltonData/issues}
}

}
\author{
\strong{Maintainer}: Zehui Yin \email{yinz39@mcmaster.ca} (\href{https://orcid.org/0000-0001-6954-7918}{ORCID})

}
\keyword{internal}
