% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Analysis}
\format{
A \emph{sf} dataframe with 49 columns, the additional columns compared to the raw grocery_DA:
\describe{
  \item{hsr_level}{The quantile level transit stop number in that census DA (high, mid, or low)}
  \item{n_hsr_stops}{The transit stop number in that census DA}
  \item{n_hsr_stops_density}{The transit stop density per square kilometres in that census DA}
  \item{pop_density}{Population density per square kilometres}
  \item{log_pop_density}{Natural log of population density per square kilometres plus 1}
  \item{log_area}{Natural log of land area in square kilometres}
  \item{dist_to_downtown}{Euclidean distance from census DA centroid to Hamilton downton in metres}
  \item{log_dist_to_downtown}{Natural log of Euclidean distance from census DA centroid to Hamilton downton in metres}
  \item{PCT_aged_under_24}{Percentage of residents aged under 24 years old}
  \item{PCT_aged_above_65}{Percentage of residents aged above 65 years old}
  \item{PCT_single_detached}{Percentage of residents living in single detached houses}
  \item{PCT_married_common_law}{Percentage of residents married or living in common law}
  \item{PCT_income_less_40k}{Percentage of residents have household income less than 40k}
  \item{PCT_income_greater_100k}{Percentage of residents have household income greater than 100k}
  \item{PCT_dont_know_official_language}{Percentage of residents don't know/speak official languages}
  \item{PCT_not_speak_offcial_language_at_home}{Percentage of residents don't speak official languages at home}
}
}
\usage{
prepare_data()
}
\value{
a data frame based on grocery_DA with additional variables
}
\description{
This function aggregates grocery store counts and computes additional
variables based on raw census data. The function does not require any
arguments but does require internet access.
}
\examples{
grocery_DA <- prepare_data()
summary(grocery_DA[,c("Freq", "PCT_single_detached")])
}
\author{
Zehui Yin, \email{yinz39@mcmaster.ca}
}
