% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_source}
\alias{data_source}
\title{Data Source Information}
\format{
A dataframe with 3 columns:
\describe{
  \item{Name}{Data Name: the name of the dataset.}
  \item{URL}{Source (URL): the link where the data is downloaded.}
  \item{Date}{Accessed Date: the date when the data is downloaded.}
}
}
\description{
This is a dataframe containing the information about all the raw data in
this package
}
\examples{
data(data_source)
summary(data_source)
head(data_source)
}
\keyword{datasets}
