% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier.R
\name{multiplier}
\alias{multiplier}
\title{Multiplier Function}
\usage{
multiplier(x, y)
}
\arguments{
\item{x}{a numerical value or vector}

\item{y}{a numerical value or vector}
}
\value{
a numerical value or vector
}
\description{
Multiplying two numerical values or vectors
}
\examples{
multiplier(2, 5)
multiplier(c(2, 5), c(3, 8))
}
